103-fibonacci.c

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int a = 1, b = 2;
-	int sum;
+	long int sum;
 
 	while (sum <= 4000000)
 	{
